% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{fix2}
\alias{fix2}
\title{fix2
funcion que muestra los datos de la matriz en vector de matrices como una tabla}
\usage{
fix2(vect_mat, n_mat)
}
\arguments{
\item{vet_mat:}{vector de matrices}

\item{n_mat:}{numero de matriz que se va pintar en la tabla
uso: > fix2(mat.data, 3), para pinta la matriz 3 del vector de matrices
Note: el vector debe tener mayor o igual cantidad de matrices que las solicitadas}
}
\description{
fix2
funcion que muestra los datos de la matriz en vector de matrices como una tabla
}
