% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{get}
\alias{get}
\title{get
funcion que dada un vector de matrices retorna la matriz n_mat}
\usage{
get(vect_mat, n_mat)
}
\arguments{
\item{vet_mat:}{vector de matrices}

\item{n_mat:}{numero de matriz que se necesita
uso: > get(mat.data, 3), para obtener la matriz 3 del vector de matrices
Note: el vector debe tener mayor o igual cantidad de matrices que las solicitadas}
}
\description{
get
funcion que dada un vector de matrices retorna la matriz n_mat
}
