% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{load.vect.df}
\alias{load.vect.df}
\title{Simulate NAs
Created by:
Created on: 2020/06/23}
\usage{
load.vect.df(n_sample, n_nas, m_source = NULL)
}
\arguments{
\item{n_sample}{number of sample}

\item{n_nas}{number of NAs in sample}

\item{m_source}{Optional matrix soruce}
}
\value{
n_sample matrix with n_nas values inside
}
\description{
Make a simulated MCAR
}
\examples{
mat.sample <- sim.load.vect.df(100,10)

}
