% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.data.R
\name{na.count}
\alias{na.count}
\title{na.count
funcion que indica la cantidad de NAs en una matriz de un vector de matrices}
\usage{
na.count(vect_mat, n_mat)
}
\arguments{
\item{vet_mat:}{vector de matrices}

\item{n_mat:}{numero de matriz para contarle los NA
uso: > na.count(mat.data, 3), Indica la cantidad de NAs en la matriz 3
Note: el vector debe tener mayor o igual cantidad de matrices que las solicitadas}
}
\description{
na.count
funcion que indica la cantidad de NAs en una matriz de un vector de matrices
}
